refactor: 코드 리뷰 반영- Business Exception 적용
