+++ src/main/java/co/kr/allpick/domain/member/sms/controller/SmsController.java
@@ -1,13 +1,18 @@
-package co.kr.allpick.domain.member.sms;
+package co.kr.allpick.domain.member.sms.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import co.kr.allpick.domain.member.service.SmsService;
+import co.kr.allpick.domain.member.sms.controller.docs.SmsControllerDocs;
+import co.kr.allpick.domain.member.sms.dto.SmsRequestDto;
+import co.kr.allpick.domain.member.sms.dto.SmsVerifyRequestDto;
+import co.kr.allpick.global.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import co.kr.allpick.global.response.ApiResponse;
-import co.kr.allpick.domain.member.service.SmsService;
 
 @RestController
 @RequestMapping("/api/sms")
@@ -16,7 +21,7 @@ public class SmsController implements SmsControllerDocs {
 
     private final SmsService smsService;
 
-    
+    @Override
     @PostMapping("/send")
     public ResponseEntity<ApiResponse<Void>> sendCode(
             @Valid @RequestBody SmsRequestDto dto) {
@@ -24,6 +29,7 @@ public class SmsController implements SmsControllerDocs {
         return ApiResponse.success("인증번호가 발송되었습니다.");
     }
 
+    @Override
     @PostMapping("/verify")
     public ResponseEntity<ApiResponse<Void>> verifyCode(
             @Valid @RequestBody SmsVerifyRequestDto dto) {
